+++ src/server/data/business/interfaces/common/IRead.ts
