+++ src/server/data/business/interfaces/common/IWrite.ts
