+++ src/server/data/repository/interfaces/IWrite.ts
